Change PID to receive integers instead of floats
This way we will not have issue with overcorrection and
when lifting off the quad should be more stable.

One of issue find in last test is that quad always start lift in the same side.
I had try balance weight but not work, this change should fix.
As error was arround 0.6 and this generate a lower proportional gain but
the integral after some time turn into a big gain.

+++ src/include/pid.h
@@ -11,6 +11,6 @@ typedef struct _pid_data {
 
 void pid_init(pid_data *pid, float p_gaing, float i_gain, float d_gain, float integraded_error_limit, float time_period);
 
-float pid_update(pid_data *pid, float target, float current);
+float pid_update(pid_data *pid, int target, int current);
 
 #endif
